added ubuntu 1804

+++ scripts/20-vbox-additions.sh
@@ -4,7 +4,14 @@ if [ ! -f  VBoxGuestAdditions.iso ]
     curl -o  VBoxGuestAdditions.iso http://download.virtualbox.org/virtualbox/${vbox_version}/VBoxGuestAdditions_${vbox_version}.iso 
 fi
 mount -o loop VBoxGuestAdditions.iso /mnt
-yum install -y kernel-headers kernel-devel bzip2 gcc
+
+if [ -f /etc/redhat-release ]; then
+  yum install -y kernel-headers kernel-devel bzip2 gcc
+fi
+
+if [ -f /etc/lsb-release ]; then
+  apt-get install -y gcc make perl
+fi
 # Install the drivers
 /mnt/VBoxLinuxAdditions.run
